Support TickTock TCP putPlain and writeLine
