+++ ticktock/src/main/java/cn/edu/tsinghua/iotdb/benchmark/ticktock/TickTockHttpWriteLine.java
@@ -20,11 +20,9 @@
 package cn.edu.tsinghua.iotdb.benchmark.ticktock;
 
 import cn.edu.tsinghua.iotdb.benchmark.measurement.Status;
-import cn.edu.tsinghua.iotdb.benchmark.schema.BaseDataSchema;
 import cn.edu.tsinghua.iotdb.benchmark.schema.DeviceSchema;
 import cn.edu.tsinghua.iotdb.benchmark.tsdb.DBConfig;
 import cn.edu.tsinghua.iotdb.benchmark.tsdb.IDatabase;
-import cn.edu.tsinghua.iotdb.benchmark.tsdb.TsdbException;
 import cn.edu.tsinghua.iotdb.benchmark.workload.ingestion.Batch;
 import cn.edu.tsinghua.iotdb.benchmark.workload.ingestion.Record;
 import org.slf4j.Logger;
@@ -33,12 +31,10 @@ import org.slf4j.LoggerFactory;
 import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 
 public class TickTockHttpWriteLine extends TickTockHttpPutPlain implements IDatabase {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(TickTockHttpWriteLine.class);
-  private static final BaseDataSchema baseDataSchema = BaseDataSchema.getInstance();
 
   /** constructor. */
   public TickTockHttpWriteLine(DBConfig dbConfig) {
@@ -49,21 +45,9 @@ public class TickTockHttpWriteLine extends TickTockHttpPutPlain implements IData
   }
 
   @Override
-  public void init() throws TsdbException {
-    HttpRequest.init();
-  }
-
-  @Override
-  public void cleanup() throws TsdbException {}
-
-  // no need for ticktock and opentsdb
-  @Override
-  public void registerSchema(List<DeviceSchema> schemaList) throws TsdbException {}
-
-  @Override
   public Status insertOneBatch(Batch batch) {
     try {
-      LinkedList<InfluxDBModel> influxDBModels = createDataModelByBatch(batch);
+      LinkedList<InfluxDBModel> influxDBModels = createInfluxDBModelByBatch(batch);
       List<String> lines = new ArrayList<>();
       for (InfluxDBModel influxDBModel : influxDBModels) {
         lines.add(model2write(influxDBModel));
@@ -81,62 +65,7 @@ public class TickTockHttpWriteLine extends TickTockHttpPutPlain implements IData
     return insertOneBatch(batch);
   }
 
-  private String model2write(InfluxDBModel influxDBModel) {
-    StringBuffer result = new StringBuffer(influxDBModel.getMetric());
-    if (influxDBModel.getTags() != null) {
-      for (Map.Entry<String, String> pair : influxDBModel.getTags().entrySet()) {
-        result.append(",");
-        result.append(pair.getKey());
-        result.append("=");
-        result.append(pair.getValue());
-      }
-    }
-    result.append(" ");
-    if (influxDBModel.getFields() != null) {
-      boolean first = true;
-      for (Map.Entry<String, Object> pair : influxDBModel.getFields().entrySet()) {
-        if (first) {
-          first = false;
-        } else {
-          result.append(",");
-        }
-        result.append(pair.getKey());
-        result.append("=");
-        // get value
-        String type =
-            typeMap(
-                baseDataSchema.getSensorType(influxDBModel.getTags().get("device"), pair.getKey()));
-        switch (type) {
-          case "BOOLEAN":
-            result.append(((boolean) pair.getValue()) ? "true" : "false");
-            break;
-          case "INT32":
-            result.append((int) pair.getValue());
-            break;
-          case "INT64":
-            result.append((long) pair.getValue());
-            break;
-          case "FLOAT":
-            result.append((float) pair.getValue());
-            break;
-          case "DOUBLE":
-            result.append((double) pair.getValue());
-            break;
-          case "TEXT":
-            result.append("\"").append(pair.getValue()).append("\"");
-            break;
-          default:
-            LOGGER.error("Unsupported data type {}, use default data type: BINARY.", type);
-            return "TEXT";
-        }
-      }
-    }
-    result.append(" ");
-    result.append(influxDBModel.getTimestamp());
-    return result.toString();
-  }
-
-  private LinkedList<InfluxDBModel> createDataModelByBatch(Batch batch) {
+  static LinkedList<InfluxDBModel> createInfluxDBModelByBatch(Batch batch) {
     DeviceSchema deviceSchema = batch.getDeviceSchema();
     List<Record> records = batch.getRecords();
     List<String> sensors = deviceSchema.getSensors();
@@ -150,7 +79,7 @@ public class TickTockHttpWriteLine extends TickTockHttpPutPlain implements IData
     return models;
   }
 
-  private InfluxDBModel createModel(
+  static InfluxDBModel createModel(
       String metric, String device, Record record, List<String> sensors) {
     InfluxDBModel model = new InfluxDBModel();
     model.setMetric(metric);
